Revisión de archivos, falta implementar en el main

+++ poo/src/main/java/com/poo/Estudiante.java
@@ -1,14 +1,14 @@
 package com.poo;
 
 public class Estudiante extends Persona {
-    private Double codigo;
+    private Integer codigo;
     private Programa programa;
     private Boolean activo;
     private Double promedio;
 
     public Estudiante(){}
 
-    public Estudiante(Double id, String nombres, String apellidos, String email, Double codigo, Programa programa, Boolean activo, Double promedio){
+    public Estudiante(Integer id, String nombres, String apellidos, String email, Integer codigo, Programa programa, Boolean activo, Double promedio){
         super(id, nombres, apellidos, email);
         this.codigo = codigo;
         this.programa = programa;
@@ -16,11 +16,11 @@ public class Estudiante extends Persona {
         this.promedio = promedio;
     }
 
-    public Double getCodigo(){
+    public Integer getCodigo(){
         return codigo;
     }
 
-     public void setCodigo(Double codigo){
+     public void setCodigo(Integer codigo){
         this.codigo = codigo;
      }
